Use int64 wherever "64-bit integer" is assumed (#902)
Miller assumes 64-bit integers, but in Go, the int type varies in size
depending on the architecture: 32-bit architectures have int
equivalent to int32. As a result, the supported range of integer
values is greatly reduced on 32-bit architectures compared to what is
suggested by the documentation.

This patch explicitly uses int64 wherever 64-bit integers are
assumed.

Test cases affected by the behaviour of the random generator are
updated to reflect the new values (the existing seed doesn't produce
the same behaviour since the way random values are generated has
changed).

+++ internal/pkg/bifs/bits.go
@@ -48,7 +48,7 @@ func bitcount_i_i(input1 *mlrval.Mlrval) *mlrval.Mlrval {
 	a = (a & _m08) + ((a >> 8) & _m08)
 	a = (a & _m16) + ((a >> 16) & _m16)
 	a = (a & _m32) + ((a >> 32) & _m32)
-	return mlrval.FromInt(int(a))
+	return mlrval.FromInt(int64(a))
 }
 
 var bitcount_dispositions = [mlrval.MT_DIM]UnaryFunc{
@@ -206,7 +206,7 @@ func ursh_i_ii(input1, input2 *mlrval.Mlrval) *mlrval.Mlrval {
 	var ua uint64 = uint64(input1.AcquireIntValue())
 	var ub uint64 = uint64(input2.AcquireIntValue())
 	var uc = ua >> ub
-	return mlrval.FromInt(int(uc))
+	return mlrval.FromInt(int64(uc))
 }
 
 var unsigned_right_shift_dispositions = [mlrval.MT_DIM][mlrval.MT_DIM]BinaryFunc{
