+++ internal/pkg/transformers/count_similar.go
@@ -155,7 +155,7 @@ func (tr *TransformerCountSimilar) Transform(
 			recordListForGroup := outer.Value.(*list.List)
 			// TODO: make 64-bit friendly
 			groupSize := recordListForGroup.Len()
-			mgroupSize := mlrval.FromInt(int(groupSize))
+			mgroupSize := mlrval.FromInt(int64(groupSize))
 			for inner := recordListForGroup.Front(); inner != nil; inner = inner.Next() {
 				recordAndContext := inner.Value.(*types.RecordAndContext)
 				recordAndContext.Record.PutCopy(tr.counterFieldName, mgroupSize)
